# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "訂餐系統",
        "title": "Order Food API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "/api/",
    "paths": {
...
        "/manager/menu/newoption": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "建立新選單",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "建立新選單",
                "parameters": [
                    {
                        "type": "MenuOption",
                        "description": "新選單",
                        "name": "data",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "商店",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/option": {
            "post": {
                "description": "新增商品選單",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "新增商品選單",
                "parameters": [
                    {
                        "type": "string",
                        "description": "選單選項",
                        "name": "selectionName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "必選數",
                        "name": "selectNum",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "商品選單",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.Option"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/option/{id}": {
            "put": {
                "description": "修改商品選單",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "修改商品選單",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "商品選單",
                        "name": "selectNum",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "刪除商品選單",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "刪除商品選單",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/selection": {
            "post": {
                "description": "新增選單選項",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "新增選單選項",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "選單編號",
                        "name": "optionID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "名稱",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "價格",
                        "name": "price",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "菜單",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.MenuSelection"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/selection/{id}": {
            "get": {
                "description": "刪除選單選項",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "刪除選單選項",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "修改選單選項",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "修改選單選項",
                "parameters": [
                    {
                        "type": "string",
                        "description": "名稱",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "價格",
                        "name": "price",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "結果",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/shop": {
            "get": {
                "description": "取得商店",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "取得商店",
                "parameters": [
                    {
                        "type": "string",
                        "description": "編號",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "商名",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "菜單",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/resp.Shop"
                            }
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "新增商店",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "新增商店",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "菜單",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.Shop"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/shop/{id}": {
            "put": {
                "description": "修改商店",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "修改商店",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "編號",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "店名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "結果",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "刪除商店",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "刪除商店",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "編號",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "resp.MenuSelection": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "Price": {
                    "type": "integer"
                }
            }
        },
        "resp.Option": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "SelectNum": {
                    "type": "integer"
                }
            }
        },
        "resp.Shop": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}